Enable std-database in native image (#12068)
* Enable std-database in native image

Not working ATM, breaks even BASE stdlib tests.
Culprit appears to be `sqlite-jdbc` client

* Workaround for sqlite-jdbc NI properies bug

sqlite-jdbc's NI properties file defines
`--enable-url-protocols=jar`. That's experimental and breaks completely
our `HostClassloader`. We exclude their config but include their feature
to correctly extract shared native library.

* fmt

* Towards error-free Table_Tests

* More reflect updates to make NI happy

* Segfaults in NI

* Pass all Table_Tests except for Excel_Spec

Excel_Spec leads to segfault, potentially related to reading/writing
files.

* Make sure `awt` is included in NI resource configs

Apparently `poi` requires `java.awt` and it was segfaulting in a
mysterious way when it was missing.

* fix moduledeps inconsistency

* Re-shufflings, de-duplication in resource configs

* legal

* Remove xalan bundle

* PR review

* nit

* Update table's reflect config

CI fails with a useless `java.lang.Exception` which is not reproducible
locally. Attempting to blindly add a few methods/fields, hoping that it
will fix it.

* Potentially fix Windows NI failures

* Workaround java.awt dependency via substitution

java.awt support on MacOS is still missing therefore
bundling/configuring resources for it only makes sense on Linux/Windows.
That was unsatisafactory.

This change uses native image's substitution mechanism to avoid any awt
library loading. The latter will not be functional when one attempts to
use it but our implementation does not rely on it at all.

* fmt