Quick Editing in Table Editor Widget (#12129)
Fixes #10865

The proper "tabbing" cells is implemented in AgGrid, but wasn't visible due to a bug. This PR adds only header edit handling.
"Entering" has our own implementation, as I haven't found a way to configure AgGrid to make behavior described in the task requirements.

https://github.com/user-attachments/assets/c3d1083f-a7c1-40c8-b206-f70a0e7a825f

# Important Notes
The "quick" editing uncovered a problem with our AgGrid updates: the component's refreshing was callled way too often. In this PR it is addressed by reducing reactive dependencies of rowData and columnDefs, so they don't cause updates when only value changes (on value change, we refresh AgGrid by hand: this does not stop editing).