Add Standard.Tableau to native-image (#12204)
* Add Standard.Tableau to native-image

Closes #12194.

* Fix sbt warning

```
[warn] there's a key that's not used by any other settings/tasks:
[warn]
[warn] * runtime-compiler-dump / javaModuleName
[warn]   +- /home/hubert/tmp/enso/build.sbt:3425
[warn]
[warn] note: a setting might still be used by a command; to exclude a key from this `lintUnused` check
[warn] either append it to `Global / excludeLintKeys` or call .withRank(KeyRanks.Invisible) on the key
```

* Test NI in nightlies

* Rewrite library loading for NI's Tableau

This change extracts `Tableau`'s Hyper shared library from the jar.
Unlike `Standard.Image`, had to provide a custom classloader that
manages to find requested library - HyperAPI used a custom JNA logic for
that.

* fmt

* fmt

* Add tests

* Provide more info to CI failures

* Extract native library from JNA

Seems to be causing problems on CI.
Extraction logic had to be extended to allow for:
- different prefixes (e.g, `linux-86-64` instead of `linux-86_64)
- classes being present on the prefix should be included in the thin
  jar; we cannot assume that extraction prefix is unique

* Fix Windows and MacOS builds

* Having a single name for OS would be too simple

* One more attempt to make Windows happy

* Set JNA boot dir explicitly

Alternative approach to deal with Windows/Mac failures on CI where
`jnidispatch` native library somehow could not be identified. Linux
worked just fine with the previous approach.

This change enhances custom classloader to identify the desired library.
That way we can provide the path to it, as specified in the
implementation of JNA.

* JNA uses deprecated extensions for native libs

`s/jnilib/dylib`

* Partial revert of #12197

Breaks Tableau integration to accomadate for opencv non-standard
packaging.